% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_dataframes_to_excel}
\alias{write_dataframes_to_excel}
\title{Write Dataframes to Excel}
\usage{
write_dataframes_to_excel(df_list, filename)
}
\arguments{
\item{df_list}{A named list of dataframes to write to the Excel file.}

\item{filename}{The name of the Excel file to create.}
}
\value{
No return value. The function prints a message indicating
the completion of the Excel file writing.
}
\description{
This function creates an Excel file with each dataframe in a list
as a separate sheet.
}
\examples{
\donttest{
tox_dat <- extr_comptox("50-00-0")
temp_file <- tempfile(fileext = ".xlsx")
write_dataframes_to_excel(tox_dat, filename = temp_file)
}
}
